Fixed Folder structure for invoice page
